requirements.txt
Pillow==7.1.2
responder==2.0.5
tensorflow-cpu==2.1.0